WIP 10/13/20: rework model and schema, pull out redundant query methods, make skeletons for etl
