+++ therapy/etl.py
@@ -1,0 +1,3 @@
+"""This module contains functions to extract data sources, transform their
+contents, and load them into the application database.
+"""
