+++ therapy/etl/chembl.py
@@ -1,0 +1,62 @@
+"""This module defines the ChEMBL normalizer"""
+from .base import Base
+from therapy import PROJECT_ROOT
+from ftplib import FTP
+import logging
+import sqlite3
+import tarfile
+
+logger = logging.getLogger('therapy')
+logger.setLevel(logging.DEBUG)
+
+
+class ChEMBL(Base):
+    """A normalizer using the ChEMBL resource."""
+
+    def _extract_data(self, *args, **kwargs):
+        if 'data_path' in kwargs:
+            chembl_db = kwargs['data_path']
+        else:
+            chembl_db = PROJECT_ROOT / 'data' / 'chembl_27' \
+                / 'chembl_27_sqlite' / 'chembl_27.db'
+        if not chembl_db.exists():
+            chembl_archive = PROJECT_ROOT / 'data' / 'chembl_27_sqlite.tar.gz'
+            chembl_archive.parent.mkdir(exist_ok=True)
+            self._download_chembl_27(chembl_archive)
+            tar = tarfile.open(chembl_archive)
+            tar.extractall()
+            tar.close()
+        assert chembl_db.exists()
+        conn = sqlite3.connect(chembl_db, check_same_thread=False)
+        self._conn = conn
+        self._cursor = conn.cursor()
+        self._create_lower_index_if_not_exists('molecule_dictionary',
+                                               'pref_name')
+        self._create_lower_index_if_not_exists('molecule_synonyms',
+                                               'synonyms')
+
+    def _create_lower_index_if_not_exists(self, table, field):
+        try:
+            command = f"""
+                CREATE INDEX lower_{field} ON {table} (lower({field}));
+            """
+            self._cursor.execute(command)
+        except sqlite3.OperationalError as e:
+            if str(e) == f'index lower_{field} already exists':
+                return False
+            else:
+                raise e
+        return True
+
+    @staticmethod
+    def _download_chembl_27(filepath):
+        logger.info('Downloading ChEMBL v27, this will take a few minutes.')
+        try:
+            with FTP('ftp.ebi.ac.uk') as ftp:
+                ftp.login()
+                logger.debug('FTP login completed.')
+                ftp.cwd('pub/databases/chembl/ChEMBLdb/releases/chembl_27')
+                with open(filepath, 'wb') as fp:
+                    ftp.retrbinary('RETR chembl_27_sqlite.tar.gz', fp.write)
+        except TimeoutError:
+            logger.error('Connection to EBI FTP server timed out.')
