+++ therapy/models.py
@@ -14,10 +14,12 @@ class Therapy(Base):
     max_phase = Column(Integer)
     withdrawn_flag = Column(Boolean)
     trade_name = Column(String)
+    src_name = Column(String, ForeignKey('meta_data.src_name'))
 
     aliases = relationship("Alias", back_populates="record")
     other_identifiers = relationship("OtherIdentifier",
                                      back_populates="record")
+    src_meta_data = relationship("Meta", back_populates="record")
 
 
 class OtherIdentifier(Base):
@@ -45,3 +47,17 @@ class Alias(Base):
     concept_id = Column(String, ForeignKey='therapies.concept_id')
 
     record = relationship("Therapy", back_populates="aliases")
+
+
+class Meta(Base):
+    """Meta info for each source"""
+
+    __tablename__ = "meta_data"
+
+    src_name = Column(String, primary_key=True)
+    data_license = Column(String)
+    data_license_url = Column(String)
+    version = Column(String)
+    data_url = Column(String)
+
+    record = relationship("Therapy", back_populates="src_meta_data")
